wasm fixes, misc impl

+++ src/Engine/sithDebugConsole.c
@@ -170,7 +170,7 @@ int sithDebugConsole_CmdTick(stdDebugConsoleCmd *pCmd, const char *pArgStr)
         if ( newTickrate >= 100 && newTickrate <= 300)
         {
             sithNet_tickrate = newTickrate;
-            sithEvent_RegisterFunc(2, (int)sithMulti_ServerLeft, newTickrate, 1);
+            sithEvent_RegisterFunc(2, sithMulti_ServerLeft, newTickrate, 1);
             result = 1;
         }
         else
